move IRIView
